Refactoring
Moved Modal components to the modals folder
